registration&default_refactored
